+++ src/main/java/com/flytrap/rssreader/service/PostCollectService.java
@@ -3,9 +3,9 @@ package com.flytrap.rssreader.service;
 import com.flytrap.rssreader.domain.post.q.PostBulkInsertPublisher;
 import com.flytrap.rssreader.domain.post.q.PostBulkInsertQueue;
 import com.flytrap.rssreader.domain.subscribe.Subscribe;
-import com.flytrap.rssreader.infrastructure.api.RssPostParser;
-import com.flytrap.rssreader.infrastructure.api.dto.RssSubscribeResource;
-import com.flytrap.rssreader.infrastructure.api.dto.RssSubscribeResource.RssItemResource;
+import com.flytrap.rssreader.infrastructure.api.parser.RssPostParser;
+import com.flytrap.rssreader.infrastructure.api.parser.dto.RssSubscribeResource;
+import com.flytrap.rssreader.infrastructure.api.parser.dto.RssSubscribeResource.RssItemResource;
 import com.flytrap.rssreader.infrastructure.entity.post.PostEntity;
 import com.flytrap.rssreader.infrastructure.entity.subscribe.SubscribeEntity;
 import com.flytrap.rssreader.infrastructure.repository.PostEntityJpaRepository;
