+++ src/test/java/com/flytrap/rssreader/fixture/FixtureFactory.java
@@ -8,12 +8,12 @@ import com.flytrap.rssreader.fixture.FixtureFields.PostEntityFields;
 import com.flytrap.rssreader.fixture.FixtureFields.RssItemResourceFields;
 import com.flytrap.rssreader.fixture.FixtureFields.SubscribeEntityFields;
 import com.flytrap.rssreader.fixture.FixtureFields.UserResourceFields;
-import com.flytrap.rssreader.infrastructure.api.dto.RssSubscribeResource.RssItemResource;
+import com.flytrap.rssreader.infrastructure.api.parser.dto.RssSubscribeResource.RssItemResource;
 import com.flytrap.rssreader.infrastructure.api.dto.UserResource;
 import com.flytrap.rssreader.infrastructure.entity.member.MemberEntity;
 import com.flytrap.rssreader.infrastructure.entity.post.PostEntity;
 import com.flytrap.rssreader.infrastructure.entity.subscribe.SubscribeEntity;
-import com.flytrap.rssreader.presentation.dto.RssFeedData;
+import com.flytrap.rssreader.infrastructure.api.parser.dto.RssFeedData;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
