+++ src/test/java/com/flytrap/rssreader/infrastructure/api/RssCheckerTest.java
@@ -2,17 +2,15 @@ package com.flytrap.rssreader.infrastructure.api;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
 import com.flytrap.rssreader.domain.subscribe.BlogPlatform;
-import com.flytrap.rssreader.presentation.dto.RssFeedData;
+import com.flytrap.rssreader.infrastructure.api.parser.RssChecker;
+import com.flytrap.rssreader.infrastructure.api.parser.dto.RssFeedData;
 import com.flytrap.rssreader.presentation.dto.SubscribeRequest.CreateRequest;
 import java.util.Optional;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -36,7 +34,6 @@ class RssCheckerTest {
         // then
 
         SoftAssertions.assertSoftly(softAssertions -> {
-            assertTrue(true);
             assertEquals(createRequest.blogUrl(), result.orElseThrow().url());
             Assertions.assertNotNull(result.get().title());
             Assertions.assertNotNull(result.get().description());
